+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add ByteCountingDecorator to HTTP module
 - Add support of HTTP status code in HttpResponse
 - Add caching HTTP client decorator
+- Add rate limiting decorator
 
 ### Changed